+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/games/GameRepresentation.java
@@ -72,6 +72,8 @@ public class GameRepresentation {
 
   private VPSChanges vpsUpdates = new VPSChanges();
 
+  private String patchVersion;
+
   private String directB2SPath;
   private String gameFilePath;
   private String povPath;
@@ -79,6 +81,14 @@ public class GameRepresentation {
   private String resPath;
   private String pupPackPath;
   private boolean altSoundAvailable;
+
+  public String getPatchVersion() {
+    return patchVersion;
+  }
+
+  public void setPatchVersion(String patchVersion) {
+    this.patchVersion = patchVersion;
+  }
 
   public int getRating() {
     return rating;