+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/highscores/parsing/listadapters/ScoreListAdapterBase.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;