+++ RELEASE_NOTES.md
@@ -3,4 +3,7 @@
 ## Changes
 
 - **Tables / Table Overview**: Added new column **Rating**. You can directly rate the game inside the column without opening the data manager.
+- **Tables / Patching**: 
+  - Added "Patch Version" field to the data manager for those who want to track this separately.
+  - Added "Patch Version" field to the patch upload dialog.
 - **Tables / VPS Tables**: Added "Comment" column and comment input field. You can enter personal comments there, e.g. to mark the table for downloading or that you tried that table but did not like it. This comment is solely stored for you and not submitted to the VP-Spreadsheet database.